api_yamdb/api/mixins.py
from rest_framework import mixins, viewsets


class CDLViewSet(mixins.CreateModelMixin,
                 mixins.DestroyModelMixin,
                 mixins.ListModelMixin,
                 viewsets.GenericViewSet):
    ...
